consolidated add_class into hawc
