+++ hawc/apps/animal/templates/animal/animalgroup_copy_selector.html
@@ -1,8 +1,5 @@
 {% extends 'animal/experiment_detail.html' %}
 
-
-{% load add_class %}
-
 {% block content %}
 
   {% include "hawc/_copy_as_new.html" with name="animal group" notes="Select an existing animal group as a template to create a new one." %}
