+++ hawc/apps/animal/templates/animal/endpoint_copy_selector.html
@@ -1,8 +1,5 @@
 {% extends 'animal/animalgroup_detail.html' %}
 
-
-{% load add_class %}
-
 {% block content %}
 
 {% include "hawc/_copy_as_new.html" with name="endpoint" notes="Select an existing endpoint as a template to create a new one. Only inputs from the current study are available." %}
