+++ hawc/apps/animal/templates/animal/endpoint_form.html
@@ -1,6 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load add_class %}
 {% load crispy_forms_tags %}
 
 {% block content %}
