+++ hawc/apps/animal/templates/animal/endpoint_list_v2.html
@@ -1,6 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load add_class %}
 {% load url_replace %}
 
 {% block content %}
