+++ hawc/apps/animal/templates/animal/experiment_copy_selector.html
@@ -1,7 +1,5 @@
 {% extends 'study/study_detail.html' %}
 
-{% load add_class %}
-
 {% block content %}
 
   {% include "hawc/_copy_as_new.html" with name="experiment" notes="Select an existing experiment as a template to create a new one." %}
