+++ hawc/apps/eco/templates/eco/result_list.html
@@ -1,6 +1,6 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load bs4 split %}
+{% load split %}
 
 {% block content %}
 <div class="d-flex mb-2">
