+++ hawc/apps/epi/templates/epi/result_copy_selector.html
@@ -1,7 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load add_class %}
-
 {% block content %}
 
   {% include "hawc/_copy_as_new.html" with name="result" notes="Select an existing result as a template to create a new one." %}
