+++ hawc/apps/epimeta/templates/epimeta/metaresult_form.html
@@ -1,7 +1,6 @@
 {% extends 'assessment-rooted.html' %}
 
 {% load crispy_forms_tags %}
-{% load add_class %}
 
 {% block content %}
   {% crispy form %}
