+++ hawc/apps/myuser/templates/myuser/set_password.html
@@ -1,7 +1,5 @@
 {% extends "admin/base_site.html" %}
 
-{% load add_class %}
-
 {% block breadcrumbs %}
 <div class="breadcrumbs">
     <a href="{% url 'admin:index' %}">Home</a>
