+++ hawc/apps/riskofbias/templates/riskofbias/arob_copy.html
@@ -1,6 +1,4 @@
 {% extends 'assessment-rooted.html' %}
-
-{% load add_class %}
 {% load crispy_forms_tags %}
 
 {% block content %}
