+++ hawc/apps/riskofbias/templates/riskofbias/riskofbiasmetric_confirm_delete.html
@@ -1,7 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load add_class %}
-
 {% block content %}
   <h2>{{object}}</h2>
   <p class='form-text text-muted'>{{object.description|safe}}</p>
