+++ hawc/apps/riskofbias/templates/riskofbias/rob_edit.html
@@ -1,6 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load add_class %}
 {% load get_at_index %}
 {% load crispy_forms_tags %}
