+++ hawc/apps/summary/templates/summary/datapivot_form.html
@@ -1,7 +1,6 @@
 {% extends 'assessment-rooted.html' %}
 
 {% load crispy_forms_tags %}
-{% load add_class %}
 
 {% block content %}
   {% include "assessment/preferred_dose_units_widget.html" %}
