+++ hawc/templates/hawc/_formset_table_template.html
@@ -1,4 +1,3 @@
-{% load add_class %}
 
 {{ formset.non_form_errors|add_class:"alert alert-danger" }}
 {{ formset.non_field_errors|add_class:"alert alert-danger" }}
